[supplier endpoints] Add service, dto and update filename to follow backend conventions
